Reduce CPU memory footprint to fit 32 GB box; add OOM-survivable checkpoint
The 32 GB box OOM-killed during the eval at step 1000. Root cause:
dSprites was loaded as float32 (9.4 GB for 573k images). With 4 train
workers + 2 eval workers, Python's refcounts break Linux's COW-on-fork,
so each worker held its own ~9 GB copy. Eval workers spawning at step
1000 pushed past the 32 GB limit.

Changes:
- dSprites: keep imgs as uint8 in RAM (2.3 GB vs 9.4 GB), convert
  per-sample to float32 in __getitem__. Single biggest win.
- num_workers: train 4 -> 2, eval 2 -> 0. Eval runs in main process so
  step-1000 spike no longer doubles the dataset-in-RAM count.
- prefetch_factor: 4 -> 2. Bounds peak buffered batches.
- Periodic checkpoint to outputs/stage_a/latest.pt at every eval_every
  step. Survives OOM kills so we don't lose multi-hour runs.

+++ glr/data/dsprites.py
@@ -54,7 +54,10 @@ class DSpritesDataset(Dataset):
                 f"  curl -L {DSPRITES_URL} -o {path}"
             )
         data = np.load(path, allow_pickle=True, encoding="latin1")
-        imgs = data["imgs"].astype(np.float32)            # (N, 64, 64) in {0,1}
+        # Keep imgs as uint8 (0/1) — 4x smaller than float32 in RAM (2.3 GB vs 9.4 GB
+        # for the full 573k set). Each DataLoader worker would otherwise hold its
+        # own COW-broken copy at 9.4 GB and OOM a 32 GB box during eval.
+        imgs = data["imgs"].astype(np.uint8)              # (N, 64, 64) in {0,1}
         latents = data["latents_classes"].astype(np.int64)  # (N, 6)
         # latents_classes columns match DSPRITES_FACTORS
 
@@ -90,7 +93,8 @@ class DSpritesDataset(Dataset):
         return out
 
     def __getitem__(self, idx: int) -> dict[str, torch.Tensor]:
-        img = self.imgs[idx]                             # (64, 64) in [0, 1]
+        # Convert per-sample from uint8 → float32 (cheap; saves RAM at dataset scale).
+        img = self.imgs[idx].astype(np.float32)          # (64, 64) in [0, 1]
         latents = self.latents[idx]
         if self.two_view:
             rng = np.random.default_rng(np.random.randint(0, 2**31 - 1))
